+++ app/src/main/java/pl/jakubtworek/menu/dto/MenuItemRequest.java
@@ -1,16 +1,17 @@
 package pl.jakubtworek.menu.dto;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public class MenuItemRequest {
     @NotNull(message = "Name cannot be null.")
     private final String name;
     @NotNull(message = "Price cannot be null.")
-    private final int price;
+    private final BigDecimal price;
     @NotNull(message = "Menu cannot be null.")
     private final String menu;
 
-    public MenuItemRequest(final String name, final int price, final String menu) {
+    public MenuItemRequest(final String name, final BigDecimal price, final String menu) {
         this.name = name;
         this.price = price;
         this.menu = menu;
@@ -20,7 +21,7 @@ public class MenuItemRequest {
         return name;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
